% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cleanForFilename}
\alias{cleanForFilename}
\title{Replace any character not acceptable as part of a filename with underscores.}
\usage{
cleanForFilename(text, onError = NA)
}
\arguments{
\item{text}{The text to clean}

\item{onError}{Value to return on NULL input - defaults to NA.}
}
\description{
\code{cleanForFilename} converts a string to US ASCII, substituting _
for any illegal characters. It then further removes anything but
numbers, characters, dash and underscore, again substituting underscores.
Since that functionality is not available on Windows, this does not handle
accented characters - they also get replaced with underscores instead of their
unaccented variants. If the input is NULL (e.g. because \code{text} is a
field that does not exist in a parsed JSON document), returns \code{onError} instead.
}
