% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{create_task}
\alias{create_task}
\title{Create a manifest describing a task to be done.}
\usage{
create_task(requestedPerformer, inputFiles, pretty = TRUE)
}
\arguments{
\item{requestedPerformer}{A string identifying the analysis package to run}

\item{inputFiles}{A data.frame with one row per input file}

\item{pretty}{Pretty print the JSON output (default TRUE)}
}
\value{
A string containing JSON, suitable for writeLines()
}
\description{
\code{create_task} takes an algorithm ID and a data.frame of input files and generates a JSON manifest.
}
\details{
The algorithm ID is a text string naming the algorithm in the Teamplay system.
The inputFiles data.frame must have the columns Description, Filename, and MIME,
in any order, and other columns will be ignored.

The allowed values in Description are up to the algorithm.
It's also up to the algorithm which MIME types it will accept, but they should
be valid - see https://www.iana.org/assignments/media-types/media-types.xhtml
}
\examples{
input_files = data.frame(
  Description = c("mutations", "methylation"),
  Filename = c("mut_export.vcf", "sample1_methylation.csv"),
  MIME = c("text/tab-separated-values", "text/csv")
)
manifest = create_task("OUS-0001", input_files)
\dontrun{writeLines(manifest, file("MANIFEST.json"))}

}
