% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{createManifest}
\alias{createManifest}
\title{Create a manifest describing a task to be done.}
\usage{
createManifest(requestedPerformer, sampleID, encounter, files, pretty = TRUE)
}
\arguments{
\item{requestedPerformer}{A string identifying the analysis package to run}

\item{sampleID}{ID of the sample/patient to be analysed}

\item{encounter}{The timepoint or encounter the data is from}

\item{files}{A data.frame with one row per input file}

\item{pretty}{Pretty print the JSON output (default TRUE)}
}
\value{
A string containing JSON, suitable for writeLines()
}
\description{
\code{createManifest} creates a manifest from an algorithm ID, sample info, and a data.frame of input files
}
\details{
The algorithm ID is a text string naming the algorithm in the Teamplay system.
The files data.frame must have the columns Description, Filename, and MIME,
in any order, and other columns will be ignored.

The sampleID and encounter are just text fields. While they accept most unicode
characters, they will be re-used by \code{\link[=packageManifest]{packageManifest()}} to build the file name.
Since that puts strict limitations on allowed characters, it's best to stick to ASCII.

The allowed values in Description are up to the algorithm.
It's also up to the algorithm which MIME types it will accept, but they should
be valid - see https://www.iana.org/assignments/media-types/media-types.xhtml
}
\examples{
input_files = data.frame(
  Description = c("mutations", "methylation"),
  Filename = c("mut_export.vcf", "sample1_methylation.csv"),
  MIME = c("text/tab-separated-values", "text/csv")
)
manifest = createManifest("OUS-0001", "Sample-001", "EOT", input_files)

}
