% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{finalizeManifest}
\alias{finalizeManifest}
\title{Update a manifest, marking it as done}
\usage{
finalizeManifest(manifest, output_files)
}
\arguments{
\item{manifest}{An existing manifest. Can be any acceptable input to
\code{\link[=readManifest]{readManifest()}}, or a jsonlite style list.}

\item{output_files}{A data.frame describing the files created}
}
\value{
A manifest in list form, suitable for \code{\link[=packageManifest]{packageManifest()}}
}
\description{
\code{finalizeManifest} marks a manifest as done and adds a lits of output files
}
\details{
Given an existing manifest describing a task to be done, e.g. from \code{\link[=createManifest]{createManifest()}},
set the status to done, last modified time to now, and append a list of
output files. The file list is in the same format as for \code{\link[=createManifest]{createManifest()}},
so it must have columns named \code{Description}, \code{Filename} and \code{MIME}.
}
\examples{
manifest_in = system.file("extdata", "MANIFEST_request.json", package="teamplayManifest")
output_files = data.frame(
  Description = c("Survival report", "Survival table"),
  Filename = c("survival_report.pdf", "survival_table.csv"),
  MIME = c("application/pdf", "text/csv")
)
manifest = finalizeManifest(manifest_in, output_files)
\dontrun{packageManifest(manifest)}

}
