% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{finish_task}
\alias{finish_task}
\title{Update a manifest, marking it as done}
\usage{
finish_task(manifest, output_files, pretty = TRUE)
}
\arguments{
\item{manifest}{An existing manifest. Can be any acceptable input to
\code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}, or a jsonlite style list.}

\item{output_files}{A data.frame describing the files created}

\item{pretty}{Pretty-print the JSON (forwarded to toJSON)}
}
\value{
A string containing JSON, suitable for writeLines()
}
\description{
\code{finish_task} updates an existing manifest with a list of output files
}
\details{
Given an existing manifest describing a task to be done, e.g. from \code{\link[=create_task]{create_task()}},
set the status to done, last modified time to now, and append a list of
output files. The file list is in the same format as for \code{\link[=create_task]{create_task()}}.
}
\examples{
manifest_in = system.file("extdata", "MANIFEST_request.json", package="teamplayManifest")
output_files = data.frame(
  Description = c("Survival report", "Survival table"),
  Filename = c("survival_report.pdf", "survival_table.csv"),
  MIME = c("application/pdf", "text/csv")
)
manifest = finish_task(manifest_in, output_files)
\dontrun{writeLines(manifest, file("MANIFEST.json"))}

}
