% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{HL7_dateTime}
\alias{HL7_dateTime}
\title{Encode a datetime string in a valid HL7 FHIR format}
\usage{
HL7_dateTime(ts = NULL)
}
\arguments{
\item{ts}{A POSIXct value to encode (default \code{\link[base:Sys.time]{base::Sys.time()}} )}
}
\value{
A string with the datetime with second precision and local timezone
}
\description{
HL7 FHIR has a DateTime data type of varying precision. If it includes
a time, it also requires the local timezone in [+-]HH:MM format,
which is subtly different from the [+-]HHMM format from \code{\link[=strftime]{strftime()}}
and \code{\link[=format.Date]{format.Date()}}. It's also picky about the precision: it seems to
require seconds if you have time at all.

See
https://www.hl7.org/fhir/datatypes.html#dateTime
}
\details{
\code{HL7_dateTime} returns a string encoding a HL7 FHIR dateTime
}
\examples{
print(HL7_dateTime())

}
