% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{HL7_decode_attachments}
\alias{HL7_decode_attachments}
\title{Decode an input or output file list to a data.frame}
\usage{
HL7_decode_attachments(file_list)
}
\arguments{
\item{file_list}{A list describing files, typically $input or $output from \code{\link[=parse_task]{parse_task()}}}
}
\value{
A data.frame with three columns describing files
}
\description{
\code{HL7_decode_attachments} decodes a file list (from parse_task or similar) into
a data frame describing input or oputput files.
}
\details{
The inverse of \code{\link[=HL7_encode_attachments]{HL7_encode_attachments()}}, this function takes a list
of the sort returned by jsonlite and simplifies it into a data.frame describing
files. It's in the same format as that accepted by \code{\link[=create_task]{create_task()}} and
\code{\link[=finish_task]{finish_task()}}.
}
\examples{
files = data.frame(
  Description = c("Survival report", "Survival table"),
  Filename = c("survival_report.pdf", "survival_table.csv"),
  MIME = c("application/pdf", "text/csv")
)
manifest = jsonlite::fromJSON(create_task("OUS-0001", files))
files == HL7_decode_attachments(manifest$input)

}
