% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{manifest_finished}
\alias{manifest_finished}
\title{Example manifest for files to be downloaded from teamplay}
\format{
An object of class \code{list} of length 9.
}
\usage{
manifest_finished
}
\description{
This is the output from jsonlite::fromJSON() for a manifest with two
input files and two output files. You can read the original JSON from
\code{system.file("extdata", "MANIFEST_finished.json", "teamplayManifest")} .
}
\keyword{datasets}
