% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{manifest_request}
\alias{manifest_request}
\title{Example manifest for files to be uploaded to teamplay}
\format{
An object of class \code{list} of length 7.
}
\usage{
manifest_request
}
\description{
This is the output from jsonlite::fromJSON() for a manifest with two
input files. You can read the original JSON from
\code{system.file("extdata", "MANIFEST_request.json", "teamplayManifest")}.
}
\keyword{datasets}
