% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{manifestToJSON}
\alias{manifestToJSON}
\title{Convert a manifest object to JSON text}
\usage{
manifestToJSON(x, pretty = FALSE, ...)
}
\arguments{
\item{x}{Manifest to convert}

\item{pretty}{Pretty-print JSON output}

\item{...}{passed to toJSON}
}
\description{
Converts a manifest object to HL7 FHIR compliant JSON.
\code{pretty} defaults to FALSE, but you can set it to TRUE for more readable
output, or to a number to use that many spaces for indentation.

It also accepts all the other settings for \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}},
but you shouldn't have to touch them.
}
\details{
\code{manifestToJSON} converts a manifest object to JSON text
}
