% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{packageManifest}
\alias{packageManifest}
\title{Package the input or output files in a manifest}
\usage{
packageManifest(manifest, fileType = "zip")
}
\arguments{
\item{manifest}{A manifest object, path to a manifest file, or JSON string}

\item{fileType}{The kind of package to create. For now the default "zip" is the
only valid choice.}
}
\description{
Depending on the \code{status} field of a manifest, this packages the files
in the \code{input} or \code{output} table, plus a JSON encoding of the manifest,
into a package. For now, the only supported \code{fileType} is zip, though there
is no reason this could not be extended to better compression methods
in later revisions.
}
\details{
\code{packageManifest} packages the input/output files (as appropriate) and the manifest to a zip file
}
