% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{parse_task}
\alias{parse_task}
\title{Parse a manifest file into something more usable}
\usage{
parse_task(manifest)
}
\arguments{
\item{manifest}{A filename or JSON string}
}
\value{
A list. The optional fields \code{input} and \code{output} are data.frames.
}
\description{
\code{parse_task} parses a JSON manifest into an R list
}
\details{
Parses a manifest into something more usable. The most interesting
fields are probably \code{input} and \code{output}, data frames describing
the input and (optionally) output files in this task.
\code{status} is either "requested" or "done", and the latter should correspond
to the existence of an \code{output} table.
}
\examples{
task = parse_task(system.file("extdata", "MANIFEST_finished.json", package="teamplayManifest"))
str(task$output)

}
