% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{printf}
\alias{printf}
\title{Print a formatted string}
\usage{
printf(fmt, ...)
}
\arguments{
\item{fmt}{a character vector of format strings, each of up to 8192 bytes.}

\item{...}{values to be passed into \code{fmt}.  Only logical,
    integer, real and character vectors are supported, but some coercion
    will be done: see the \sQuote{Details} section. Up to 100.}
}
\description{
For some reason R has \code{\link[=sprintf]{sprintf()}} for formatting strings, and \code{\link[=cat]{cat()}} for
printing untouched strings, but not a classic printf for printing
untouched, formatted strings. This function is literally \code{cat(sprintf(fmt, ...))}.
}
\details{
\code{printf} prints a formatted string in the style of sprintf
}
\examples{
printf("pi=\%1.2f", pi);  printf(", and \%d in hex is \%x\n", 100, 100)
}
