% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{readManifest}
\alias{readManifest}
\title{Read a manifest from a file}
\usage{
readManifest(manifest)
}
\arguments{
\item{manifest}{A filename or JSON string}
}
\value{
A list. The optional fields \code{input} and \code{output} are data.frames.
}
\description{
\code{readManifest} parses a JSON manifest into a manifest object
}
\details{
Parses a manifest into something more usable.
The manifest class is a list with some helper functions. The most interesting
fields are probably \code{input} and \code{output}, data frames describing
the input and (optionally) output files in this task.
\code{status} is either "requested" or "done", and the latter should correspond
to the existence of an \code{output} table.
}
\examples{
task = readManifest(system.file("extdata", "MANIFEST_finished.json", package="teamplayManifest"))
str(task$output)

}
