% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{HL7_decode_attachments}
\alias{HL7_decode_attachments}
\title{Decode an input or output file list to a data.frame}
\usage{
HL7_decode_attachments(file_list)
}
\arguments{
\item{file_list}{A list describing files, typically $input or $output from
reading a manifest with \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}.}
}
\value{
A data.frame with three columns describing files
}
\description{
\code{HL7_decode_attachments} decodes a file list (from parse_task or similar) into
a data frame describing input or oputput files.
}
\details{
The inverse of \code{\link[=HL7_encode_attachments]{HL7_encode_attachments()}}, this function takes a list
of the sort returned by jsonlite and simplifies it into a data.frame describing
files. The returned data.frame is in the same format as that accepted by
\code{\link[=createManifest]{createManifest()}} and \code{\link[=finalizeManifest]{finalizeManifest()}}.
}
\examples{
files = data.frame(
  Description = c("Survival report", "Survival table"),
  Filename = c("survival_report.pdf", "survival_table.csv"),
  MIME = c("application/pdf", "text/csv")
)
manifest = createManifest("OUS-0001", "Sample1", "T1", files)
files == HL7_decode_attachments(jsonlite::fromJSON(manifestToJSON(manifest))$input)
files == readManifest(manifestToJSON(manifest))$input

}
